add player hint tracking and basic early game stat
Basically, now the player will hint at a 1 during early game (< 1/2 of
the board is empty).

It will do a color hint if it hasn't done it before, and if the 1 is the
only color in the deck.

It will do a number hint if it doesn't do a color hint and it hasn't
done it before.

Else it will blindly discard from the right.